requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
langchain_community==0.0.29
langchain-groq==0.1.0
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.6.1
setuptools>=65.5.0
uvloop>=0.17.0
httptools>=0.5.0